chore: sets up mantine for UI lib
