+++ package.json
@@ -3,9 +3,14 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
+		"@emotion/react": "^11.10.5",
+		"@emotion/styled": "^11.10.5",
+		"@mantine/core": "^5.9.4",
+		"@mantine/hooks": "^5.9.4",
 		"@testing-library/jest-dom": "^5.16.4",
 		"@testing-library/react": "^13.1.1",
 		"@testing-library/user-event": "^13.5.0",
+		"framer-motion": "^6.5.1",
 		"react": "^18.0.0",
 		"react-dom": "^18.0.0",
 		"react-scripts": "5.0.1",
@@ -41,7 +46,6 @@
 	"devDependencies": {
 		"@commitlint/cli": "^17.3.0",
 		"@commitlint/config-conventional": "^17.3.0",
-		"autoprefixer": "^10.4.13",
 		"eslint": "^8.29.0",
 		"eslint-config-airbnb": "^19.0.4",
 		"eslint-config-prettier": "^8.5.0",
@@ -51,8 +55,6 @@
 		"eslint-plugin-react": "^7.31.11",
 		"eslint-plugin-react-hooks": "^4.6.0",
 		"husky": "^8.0.2",
-		"postcss": "^8.4.20",
-		"prettier": "^2.8.1",
-		"tailwindcss": "^3.2.4"
+		"prettier": "^2.8.1"
 	}
 }
