--- src/App.js
@@ -1,11 +1,0 @@
-import './App.css';
-
-function App() {
-	return (
-		<div>
-			<h1 className="text-2xl font-bold">Hi, from Tailwind</h1>
-		</div>
-	);
-}
-
-export default App;
