+++ src/App.jsx
@@ -1,0 +1,12 @@
+import { Button, Title } from '@mantine/core';
+
+function App() {
+	return (
+		<div>
+			<Title order={1}>This is h1 title</Title>
+			<Button variant="filled">Click me, I am a Mantine Button!</Button>
+		</div>
+	);
+}
+
+export default App;
