+++ src/index.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { MantineProvider } from '@mantine/core';
+
+import App from './App';
+
+import './index.css';
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+	<React.StrictMode>
+		<MantineProvider withGlobalStyles withNormalizeCSS>
+			<App />
+		</MantineProvider>
+	</React.StrictMode>
+);
